requirements.txt
fastapi 
uvicorn
pymongo
langchain
python-dotenv